+++ app/src/main/java/net/sigmabeta/chipbox/ui/games/GameGridPresenter.kt
@@ -29,7 +29,7 @@ class GameGridPresenter @Inject constructor(val database: SongDatabaseHelper) : 
      */
 
     override fun setup(arguments: Bundle?) {
-        platform = arguments?.getInt(GameGridFragment.ARGUMENT_PLATFORM_INDEX) ?: Track.PLATFORM_UNDEFINED
+        platform = arguments?.getLong(GameGridFragment.ARGUMENT_PLATFORM_INDEX) ?: Track.PLATFORM_UNDEFINED
 
         val subscription = database.getGamesList(platform)
                 .subscribeOn(Schedulers.io())
